multi threaded

+++ java/src/main/java/at/bumzack/fractalthingi/Hsl.java
@@ -2,7 +2,7 @@ package at.bumzack.fractalthingi;
 
 
 public class Hsl {
-    int h;
-    int s;
-    int l;
+    double h;
+    double s;
+    double l;
 }
